refactor: 도입 BaseMastermindSolver 및 솔버 3종 상속 구조 리팩토링 (#38)
