+++ src/solvers/entropy_solver_old.py
